JDK-8241840 The tests are now reused for native and monocle tests.
